tota

+++ Assignment1_Deeandra_Hernandez.c
@@ -194,7 +194,7 @@ double calculateWidth()
 
 double calculateHeight()
 {
- int x1,y1,x2,y2;
+ double x1,y1,x2,y2;
 
  Pointsprint(&x1, &y1, &x2, &y2);
